feat: added in memory cache w\o debug chore: updated golangci-lint config and Taskfile

+++ internal/pkg/server/server.go
@@ -6,8 +6,8 @@ import (
 )
 
 type Config struct {
-	HTTPPort        int `mapstructure:"http_port"`
-	ShutdownTimeout int `mapstructure:"shutdown_timeout"`
+	HTTPPort        int16 `mapstructure:"http_port"`
+	ShutdownTimeout int16 `mapstructure:"shutdown_timeout"`
 }
 
 func New(c *Config) *http.Server {
